Scheduler

MiniEvents.jl

EventLists

Index

MiniEvents

Scheduler

Base.isempty — Method

isempty(scheduler)

Returns true if the scheduler does not contain any actions.

MiniEvents.Scheduler.advance! — Method

advance!(scheduler, t)

Advance time to t. Caution, this does not check for consistency!

MiniEvents.Scheduler.next! — Method

next!(scheduler)

Run the next action in scheduler or do nothing if empty. Returns the action's return value.

MiniEvents.Scheduler.reset! — Method

reset!(scheduler)

Remove all actions from scheduler and reset time to 0.

MiniEvents.Scheduler.schedule_at! — Method

schedule_at!(fun, obj, at, scheduler)

Add a single item to the scheduler. Adds function fun to be called on obj at time at to scheduler.

MiniEvents.Scheduler.schedule_in! — Method

schedule_in!(fun, obj, wait, scheduler)

Add a single item (fun to be called on obj) at wait time from now to scheduler.

MiniEvents.Scheduler.time_next — Method

time_next(scheduler)

Time stamp of the next action to be executed by scheduler or time_now if it is empty.

MiniEvents.Scheduler.time_now — Method

time_now(scheduler)

Time stamp of the last action that was executed by scheduler.

MiniEvents.Scheduler.unschedule! — Method

unschedule!(scheduler, obj)

Remove action for obj from scheduler.

MiniEvents.Scheduler.upto! — Method

upto!(scheduler, atime)

Run actions in scheduler up to time atime. Returns the scheduler.

MiniEvents.Scheduler.PQScheduler — Type

A simple scheduler based on PriorityQueue.

MiniEvents.Scheduler.PQScheduler — Method

Construct an empty PQScheduler with a give TIME type.

Index

MiniEvents.EventLists.AgentEvents

MiniEvents.EventLists.EventList

MiniEvents.EventLists.EventList

MiniEvents.Scheduler.PQScheduler

MiniEvents.Scheduler.PQScheduler

Base.isempty

MiniEvents.EventLists.add_agent!

MiniEvents.EventLists.change_rates!

MiniEvents.EventLists.remove_agent!

MiniEvents.EventLists.sum_rates

MiniEvents.EventLists.sum_rates

MiniEvents.Scheduler.advance!

MiniEvents.Scheduler.next!

MiniEvents.Scheduler.reset!

MiniEvents.Scheduler.schedule_at!

MiniEvents.Scheduler.schedule_in!

MiniEvents.Scheduler.time_next

MiniEvents.Scheduler.time_now

MiniEvents.Scheduler.unschedule!

MiniEvents.Scheduler.upto!

MiniEvents.kill!

MiniEvents.next_event!

MiniEvents.now

MiniEvents.refresh!

MiniEvents.schedule!

MiniEvents.schedule_dt!

MiniEvents.spawn!

MiniEvents.spawn_pop!

MiniEvents.step_dt!

MiniEvents.step_until!

MiniEvents.@events

MiniEvents.@simulation